requirements.txt
fire
loguru
manga-ocr
natsort
numpy==1.26
opencv-python
Pillow
pillow-avif-plugin
pyclipper
pytest
requests
scipy
shapely
torchsummary
tqdm
uuid-utils
yattag
